+++ NEWS.md
@@ -3,6 +3,7 @@ __kumu 0.0.0.9000 (in development)__
 
 ### NEW FEATURES
 
+ * Adds basic workflow to interact with Tetrad CMD interface [#2](https://github.com/sailuh/kumu/issues/2) 
  * Defines basic package structure [#1](https://github.com/sailuh/kumu/issues/1)
 
 ### MINOR IMPROVEMENTS